+++ .eslintrc.json
@@ -36,6 +36,7 @@
               "ClassDeclaration": true
           }
       }],
+      "react/forbid-prop-types": 0,
       "jsx-a11y/anchor-is-valid": [ "error", { "components": [ "Link" ], "specialLink": [ "to" ] } ]
     }
   }
